compile again on Windows. Doesn't run for me though....
[git-p4: depot-paths = "//depot/qt/main/": change = 127237]

+++ src/gui/text/qfont_win.cpp
@@ -379,7 +379,7 @@ int QFontMetrics::width( QChar ch ) const
     int nglyphs = 7;
     engine->stringToCMap( &ch, 1, glyphs, &nglyphs, 0 );
 
-    return glyphs[0].advance.x;
+    return glyphs[0].advance.x.toInt();
 }
 
 
@@ -390,7 +390,7 @@ int QFontMetrics::charWidth( const QString &str, int pos ) const
 
     QTextEngine layout( str,  d );
     layout.itemize( QTextEngine::WidthOnly );
-    int w = layout.width( pos, 1 );
+    int w = layout.width( pos, 1 ).toInt();
 
     if ( (qWinVersion() & Qt::WV_NT_based) == 0 ) {
 	QFontEngine *engine = d->engineForScript( (QFont::Script) fscript );
